refactor(condo): DOMA-4356 extract property map edit forms to detached components
