+++ apps/condo/domains/property/components/panels/Builder/BuildingPanelEdit.tsx
@@ -1,29 +1,26 @@
 /** @jsx jsx */
-import { CloseOutlined, DeleteFilled, DownOutlined } from '@ant-design/icons'
+import { CloseOutlined } from '@ant-design/icons'
 import { BuildingMap, BuildingSection, BuildingUnit, BuildingUnitSubType, Property as PropertyType } from '@app/condo/schema'
 import { Button } from '@condo/domains/common/components/Button'
 import { colors, fontSizes, shadows } from '@condo/domains/common/constants/style'
 import { UnitButton } from '@condo/domains/property/components/panels/Builder/UnitButton'
-import { MAX_PROPERTY_FLOORS_COUNT, MAX_PROPERTY_UNITS_COUNT_PER_FLOOR, MIN_SECTIONS_TO_SHOW_FILTER } from '@condo/domains/property/constants/property'
+import { MIN_SECTIONS_TO_SHOW_FILTER } from '@condo/domains/property/constants/property'
 import { Property } from '@condo/domains/property/utils/clientSchema'
 import { useIntl } from '@condo/next/intl'
 import { css, jsx } from '@emotion/react'
 import styled from '@emotion/styled'
 import {
     Col,
-    InputNumber,
     notification,
     Row,
     RowProps,
     Space,
     Typography,
 } from 'antd'
-import Input from '@condo/domains/common/components/antd/Input'
 import Select from '@condo/domains/common/components/antd/Select'
 import cloneDeep from 'lodash/cloneDeep'
 import debounce from 'lodash/debounce'
 import get from 'lodash/get'
-import isEmpty from 'lodash/isEmpty'
 import isFunction from 'lodash/isFunction'
 import isNull from 'lodash/isNull'
 import last from 'lodash/last'
@@ -42,14 +39,15 @@ import {
     MapSectionContainer, PropertyMapFloor,
     UnitTypeLegendItem,
 } from './BuildingPanelCommon'
+import { UnitForm } from '@condo/domains/property/components/panels/Builder/forms/UnitForm'
+import { AddSectionForm, EditSectionForm } from '@condo/domains/property/components/panels/Builder/forms/SectionForm'
+import { AddSectionFloor } from '@condo/domains/property/components/panels/Builder/forms/FloorForm'
+import { AddParkingForm, EditParkingForm } from '@condo/domains/property/components/panels/Builder/forms/ParkingForm'
+import { ParkingUnitForm } from '@condo/domains/property/components/panels/Builder/forms/ParkingUnitForm'
 
 import { FullscreenHeader, FullscreenWrapper } from './Fullscreen'
 import { MapEdit, MapViewMode } from './MapConstructor'
-import Checkbox from '@condo/domains/common/components/antd/Checkbox'
 
-const { Option } = Select
-
-const INPUT_STYLE = { width: '100%' }
 const DEBOUNCE_TIMEOUT = 800
 const INSTANT_ACTIONS = ['addBasement', 'addAttic']
 
@@ -112,13 +110,6 @@ const TopModal = styled.div<ITopModalProps>`
   }
 `
 
-const FormModalCss = css`
-  & .ant-space,
-  & button {
-    width: 100%;
-  }
-`
-
 export const AddressTopTextContainer = styled.div`
   font-size: ${fontSizes.content};
   line-height: 24px;
@@ -173,7 +164,7 @@ const useHotkeyToSaveProperty = (map, mapEdit, property) => {
     const intl = useIntl()
     const ChangesSaved = intl.formatMessage({ id: 'ChangesSaved' })
     const MapValidationError = intl.formatMessage({ id: 'pages.condo.property.warning.modal.SameUnitNamesErrorMsg' })
-    
+
     const quickSave = Property.useUpdate({}, () => notification.success({
         message: ChangesSaved,
         placement: 'bottomRight',
@@ -690,979 +681,3 @@ const PropertyMapUnit: React.FC<IPropertyMapUnitProps> = ({ builder, refresh, un
         >{unit.label}</UnitButton>
     )
 }
-
-
-interface IPropertyMapModalForm {
-    builder: MapEdit
-    refresh(): void
-}
-const MODAL_FORM_ROW_GUTTER: RowProps['gutter'] = [0, 24]
-const MODAL_FORM_ROW_BUTTONS_GUTTER: RowProps['gutter'] = [0, 16]
-const MODAL_FORM_BUTTON_STYLE: React.CSSProperties = { marginTop: '12px' }
-const MODAL_FORM_CHECKBOX_STYLE: React.CSSProperties = { marginTop: '20px' }
-
-const AddSectionForm: React.FC<IPropertyMapModalForm> = ({ builder, refresh }) => {
-    const intl = useIntl()
-    const SectionNameLabel = intl.formatMessage({ id: 'pages.condo.property.section.form.numberOfSection' })
-    const MinFloorLabel = intl.formatMessage({ id: 'pages.condo.property.section.form.minfloor' })
-    const FloorCountLabel = intl.formatMessage({ id: 'pages.condo.property.section.form.floorCount' })
-    const UnitsOnFloorLabel = intl.formatMessage({ id: 'pages.condo.property.section.form.unitsOnFloor' })
-    const CreateNewLabel = intl.formatMessage({ id: 'pages.condo.property.section.form.mode.create' })
-    const UnitTypeLabel = intl.formatMessage({ id: 'pages.condo.property.modal.UnitType' })
-    const CopyLabel = intl.formatMessage({ id: 'pages.condo.property.section.form.mode.copy' })
-    const AddLabel = intl.formatMessage({ id: 'Add' })
-    const ShowMinFloor = intl.formatMessage({ id: 'pages.condo.property.parking.form.showMinFloor' })
-    const HideMinFloor = intl.formatMessage({ id: 'pages.condo.property.parking.form.hideMinFloor' })
-
-    const [minFloor, setMinFloor] = useState(1)
-    const [floorCount, setFloorCount] = useState(null)
-    const [unitsOnFloor, setUnitsOnFloor] = useState(null)
-    const [copyId, setCopyId] = useState<string | null>(null)
-    const [minFloorHidden, setMinFloorHidden] = useState<boolean>(true)
-    const [sectionName, setSectionName] = useState<string>(builder.nextSectionName)
-    const [unitType, setUnitType] = useState<BuildingUnitSubType>(BuildingUnitSubType.Flat)
-
-    const resetForm = useCallback(() => {
-        setMinFloor(1)
-        setFloorCount(null)
-        setUnitsOnFloor(null)
-    }, [])
-
-    const toggleMinFloorVisible = useCallback(() => {
-        if (!minFloorHidden) {
-            setMinFloor(1)
-        }
-        setMinFloorHidden(!minFloorHidden)
-    }, [minFloorHidden])
-
-    const setMinFloorValue = useCallback((value) => { setMinFloor(value) }, [])
-    const setFloorCountValue = useCallback((value) => { setFloorCount(value) }, [])
-    const maxFloorValue = useMemo(() => {
-        if (floorCount === 1) return minFloor
-        if (minFloor > 0) return floorCount + minFloor - 1
-        return floorCount + minFloor
-    }, [floorCount, minFloor])
-
-    useEffect(() => {
-        if (minFloor && floorCount && unitsOnFloor && sectionName) {
-            builder.addPreviewSection({
-                id: '',
-                name: sectionName,
-                minFloor,
-                maxFloor: maxFloorValue,
-                unitsOnFloor,
-            }, unitType)
-            refresh()
-        } else {
-            builder.removePreviewSection()
-        }
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [minFloor, floorCount, unitsOnFloor, sectionName, unitType])
-
-    useEffect(() => {
-        if (copyId !== null) {
-            builder.addPreviewCopySection(copyId)
-        }
-    }, [builder, copyId])
-
-    const handleFinish = useCallback(() => {
-        builder.removePreviewSection()
-        if (copyId === null) {
-            builder.addSection({ id: '', name: sectionName, minFloor, maxFloor: maxFloorValue, unitsOnFloor }, unitType)
-            setSectionName(builder.nextSectionName)
-        } else {
-            builder.addCopySection(copyId)
-        }
-
-        refresh()
-        resetForm()
-    }, [refresh, resetForm, builder, sectionName, minFloor, unitsOnFloor, unitType, copyId, maxFloorValue])
-
-    const setSectionNameValue = useCallback((value) => setSectionName(value ? value.toString() : ''), [])
-    const isSubmitDisabled = copyId !== null ? false : !(minFloor && floorCount && unitsOnFloor)
-    const isCreateColumnsHidden = copyId !== null
-    const iconRotation = minFloorHidden ? 0 : 180
-    const minFloorMargin = minFloorHidden ? '-28px' : 0
-
-    return (
-        <Row gutter={MODAL_FORM_ROW_GUTTER} css={FormModalCss}>
-            <Col span={24}>
-                <Select value={copyId} onSelect={setCopyId} disabled={builder.isEmptySections}>
-                    <Select.Option key='create' value={null}>{CreateNewLabel}</Select.Option>
-                    {builder.map.sections.filter(section => !section.preview).map(section => (
-                        <Select.Option
-                            key={`copy-${section.id}`}
-                            value={section.id}
-                        >
-                            {CopyLabel}{section.name}
-                        </Select.Option>
-                    ))}
-                </Select>
-            </Col>
-            <Col span={24} hidden={isCreateColumnsHidden}>
-                <Space direction='vertical' size={8}>
-                    <Typography.Text type='secondary'>{SectionNameLabel}</Typography.Text>
-                    <InputNumber
-                        value={sectionName}
-                        min={1}
-                        onChange={setSectionNameValue}
-                        style={INPUT_STYLE}
-                        type='number'
-                    />
-                </Space>
-            </Col>
-            <Col span={24} hidden={isCreateColumnsHidden}>
-                <Space direction='vertical' size={8}>
-                    <Typography.Text type='secondary'>{FloorCountLabel}</Typography.Text>
-                    <InputNumber value={floorCount} onChange={setFloorCountValue} min={1} max={MAX_PROPERTY_FLOORS_COUNT} style={INPUT_STYLE} type='number' />
-                </Space>
-            </Col>
-            <Col span={24} hidden={isCreateColumnsHidden}>
-                <Space direction='vertical' size={8}>
-                    <Typography.Text type='secondary'>{UnitTypeLabel}</Typography.Text>
-                    <Select value={unitType} onSelect={setUnitType}>
-                        {Object.values(BuildingUnitSubType)
-                            .filter(unitType => unitType !== BuildingUnitSubType.Parking)
-                            .map((unitType, key) => (
-                                <Select.Option key={`${key}-${unitType}`} value={unitType} title={unitType}>
-                                    {intl.formatMessage({ id: `pages.condo.property.modal.unitType.${unitType}` })}
-                                </Select.Option>
-                            ))}
-                    </Select>
-                </Space>
-            </Col>
-            <Col span={24} hidden={isCreateColumnsHidden} style={{ marginTop: minFloorMargin }}>
-                <Space
-                    direction='vertical'
-                    size={8}
-                    hidden={minFloorHidden}
-                >
-                    <Typography.Text type='secondary'>{MinFloorLabel}</Typography.Text>
-                    <InputNumber
-                        value={minFloor}
-                        onChange={setMinFloorValue}
-                        style={INPUT_STYLE}
-                        type='number'
-                    />
-                </Space>
-                <Typography.Text onClick={toggleMinFloorVisible} style={TEXT_BUTTON_STYLE}>
-                    {minFloorHidden ? ShowMinFloor : HideMinFloor} <DownOutlined rotate={iconRotation}/>
-                </Typography.Text>
-            </Col>
-            <Col span={24} hidden={isCreateColumnsHidden}>
-                <Space direction='vertical' size={8}>
-                    <Typography.Text type='secondary'>{UnitsOnFloorLabel}</Typography.Text>
-                    <InputNumber min={1} max={MAX_PROPERTY_UNITS_COUNT_PER_FLOOR} value={unitsOnFloor} onChange={setUnitsOnFloor} style={INPUT_STYLE} type='number'/>
-                </Space>
-            </Col>
-            <Col span={24}>
-                <Button
-                    key='submit'
-                    secondary
-                    onClick={handleFinish}
-                    type='sberDefaultGradient'
-                    style={MODAL_FORM_BUTTON_STYLE}
-                    disabled={isSubmitDisabled}
-                > {AddLabel} </Button>
-            </Col>
-        </Row>
-    )
-}
-
-const BUTTON_SPACE_SIZE = 28
-
-const UnitForm: React.FC<IPropertyMapModalForm> = ({ builder, refresh }) => {
-    const intl = useIntl()
-    const mode = builder.editMode
-    const SaveLabel = intl.formatMessage({ id: mode === 'editUnit' ? 'Save' : 'Add' })
-    const DeleteLabel = intl.formatMessage({ id: 'Delete' })
-    const NameLabel = intl.formatMessage({ id: 'pages.condo.property.unit.Name' })
-    const SectionLabel = intl.formatMessage({ id: 'pages.condo.property.section.Name' })
-    const FloorLabel = intl.formatMessage({ id: 'pages.condo.property.floor.Name' })
-    const UnitTypeLabel = intl.formatMessage({ id: 'pages.condo.property.modal.UnitType' })
-    const RenameNextUnitsLabel = intl.formatMessage({ id: 'pages.condo.property.modal.RenameNextUnits' })
-
-    const [label, setLabel] = useState('')
-    const [floor, setFloor] = useState('')
-    const [section, setSection] = useState('')
-    const [unitType, setUnitType] = useState<BuildingUnitSubType>(BuildingUnitSubType.Flat)
-
-    const [sections, setSections] = useState([])
-    const [floors, setFloors] = useState([])
-
-    const renameNextUnits = useRef(true)
-
-    const updateSection = (value) => {
-        setSection(value)
-        setFloors(builder.getSectionFloorOptions(value))
-        if (mode === 'editUnit') {
-            const mapUnit = builder.getSelectedUnit()
-            if (value === mapUnit.section) {
-                setFloor(mapUnit.floor)
-            } else {
-                setFloor(null)
-            }
-        } else {
-            setFloor(null)
-        }
-    }
-
-    useEffect(() => {
-        setSections(builder.getSectionOptions())
-        const mapUnit = builder.getSelectedUnit()
-        if (mapUnit) {
-            setFloors(builder.getSectionFloorOptions(mapUnit.section))
-            setLabel(mapUnit.label)
-            setSection(mapUnit.section)
-            setFloor(mapUnit.floor)
-            setUnitType(mapUnit.unitType)
-        }
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [builder])
-
-    useEffect(() => {
-        if (label && floor && section && unitType && mode === 'addUnit') {
-            builder.addPreviewUnit({ id: '', label, floor, section, unitType }, renameNextUnits.current)
-            refresh()
-        } else {
-            builder.removePreviewUnit(renameNextUnits.current)
-        }
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [label, floor, section, mode, unitType])
-
-    const resetForm = useCallback(() => {
-        setLabel('')
-        setFloor('')
-        setSection('')
-    }, [])
-
-    const toggleRenameNextUnits = useCallback((event) => { renameNextUnits.current = event.target.checked }, [])
-
-    const applyChanges = useCallback(() => {
-        const mapUnit = builder.getSelectedUnit()
-        if (mapUnit) {
-            builder.updateUnit({ ...mapUnit, label, floor, section, unitType }, renameNextUnits.current)
-        } else {
-            builder.removePreviewUnit()
-            builder.addUnit({ id: '', label, floor, section, unitType }, renameNextUnits.current)
-            resetForm()
-        }
-        refresh()
-    }, [builder, refresh, resetForm, label, floor, section, unitType])
-
-    const deleteUnit = useCallback(() => {
-        const mapUnit = builder.getSelectedUnit()
-        builder.removeUnit(mapUnit.id, renameNextUnits.current)
-        refresh()
-        resetForm()
-    }, [resetForm, refresh, builder])
-
-    const updateUnitType = useCallback((value) => {
-        setUnitType(value)
-    }, [])
-
-    const isApplyButtonDisabled = useMemo(() => {
-        let isUnitLabelUnique = true
-        if (mode === 'addUnit') {
-            isUnitLabelUnique = builder.validateUniqueUnitLabel()
-        } else if (mode === 'editUnit') {
-            const selectedUnit = builder.getSelectedUnit()
-            const unitPlacementChanged = selectedUnit.floor !== floor
-                || selectedUnit.section !== section
-                || unitType !== selectedUnit.unitType
-            const labelChanged = selectedUnit.label !== label
-            const labelValidation = labelChanged
-                ? builder.validateUniqueUnitLabel(label, 'section')
-                : false
-
-            isUnitLabelUnique = unitPlacementChanged
-                ? (!labelChanged || labelValidation)
-                : labelValidation
-        }
-
-        return !(floor && section && label.trim() && isUnitLabelUnique)
-    }, [floor, section, label, unitType, builder, mode])
-
-    return (
-        <Row gutter={MODAL_FORM_ROW_GUTTER} css={FormModalCss}>
-            <Col span={24}>
-                <Space direction='vertical' size={8}>
-                    <Typography.Text>{UnitTypeLabel}</Typography.Text>
-                    <Select
-                        value={intl.formatMessage({ id: `pages.condo.property.modal.unitType.${unitType}` })}
-                        onSelect={updateUnitType}
-                        style={INPUT_STYLE}
-                    >
-                        {Object.values(BuildingUnitSubType)
-                            .filter(unitType => unitType !== BuildingUnitSubType.Parking)
-                            .map((unitType, unitTypeIndex) => (
-                                <Option key={unitTypeIndex} value={unitType}>{intl.formatMessage({ id: `pages.condo.property.modal.unitType.${unitType}` })}</Option>
-                            ))}
-                    </Select>
-                </Space>
-            </Col>
-            <Col span={24}>
-                <Space direction='vertical' size={8}>
-                    <Typography.Text type='secondary'>{NameLabel}</Typography.Text>
-                    <Input allowClear={true} value={label} onChange={e => setLabel(e.target.value)} style={INPUT_STYLE} />
-                </Space>
-            </Col>
-            <Col span={24}>
-                <Space direction='vertical' size={8} style={INPUT_STYLE}>
-                    <Typography.Text type='secondary' >{SectionLabel}</Typography.Text>
-                    <Select value={section} onSelect={updateSection} style={INPUT_STYLE}>
-                        {sections.map((sec) => {
-                            return <Option key={sec.id} value={sec.id}>{sec.label}</Option>
-                        })}
-                    </Select>
-                </Space>
-            </Col>
-            <Col span={24}>
-                <Space direction='vertical' size={BUTTON_SPACE_SIZE}>
-                    <Space direction='vertical' size={8} style={INPUT_STYLE}>
-                        <Typography.Text type='secondary' >{FloorLabel}</Typography.Text>
-                        <Select value={floor} onSelect={setFloor} style={INPUT_STYLE}>
-                            {floors.map(floorOption => {
-                                return <Option key={floorOption.id} value={floorOption.id}>{floorOption.label}</Option>
-                            })}
-                        </Select>
-                        <Checkbox defaultChecked onChange={toggleRenameNextUnits} style={MODAL_FORM_CHECKBOX_STYLE}>
-                            {RenameNextUnitsLabel}
-                        </Checkbox>
-                    </Space>
-                    <Row gutter={MODAL_FORM_ROW_BUTTONS_GUTTER}>
-                        <Col span={24}>
-                            <Button
-                                secondary
-                                onClick={applyChanges}
-                                type='sberDefaultGradient'
-                                disabled={isApplyButtonDisabled}
-                            > {SaveLabel} </Button>
-                        </Col>
-                        {
-                            mode === 'editUnit' && (
-                                <Col span={24}>
-                                    <Button
-                                        secondary
-                                        onClick={deleteUnit}
-                                        type='sberDangerGhost'
-                                        icon={<DeleteFilled />}
-                                    >{DeleteLabel}</Button>
-                                </Col>
-                            )
-                        }
-                    </Row>
-                </Space>
-            </Col>
-        </Row>
-    )
-}
-
-const MODAL_FORM_EDIT_GUTTER: RowProps['gutter'] = [0, 28]
-const MODAL_FORM_BUTTON_GUTTER: RowProps['gutter'] = [0, 16]
-
-const EditSectionForm: React.FC<IPropertyMapModalForm> = ({ builder, refresh }) => {
-    const intl = useIntl()
-    const NameLabel = intl.formatMessage({ id: 'pages.condo.property.section.form.name' })
-    const NamePlaceholderLabel = intl.formatMessage({ id: 'pages.condo.property.section.form.name.placeholder' })
-    const SaveLabel = intl.formatMessage({ id: 'Save' })
-    const DeleteLabel = intl.formatMessage({ id: 'Delete' })
-    const RenameNextUnitsLabel = intl.formatMessage({ id: 'pages.condo.property.modal.RenameNextUnits' })
-
-    const [name, setName] = useState<string>('')
-    const renameNextUnits = useRef(true)
-
-    const section = builder.getSelectedSection()
-
-    useEffect(() => {
-        setName(section ? section.name : '')
-    }, [section])
-
-    const setNameValue = useCallback((value) => setName(value ? value.toString() : ''), [])
-    const toggleRenameNextUnits = useCallback((event) => { renameNextUnits.current = event.target.checked }, [])
-
-    const updateSection = useCallback(() => {
-        builder.updateSection({ ...section, name })
-        refresh()
-    }, [builder, refresh, name, section])
-
-    const deleteSection = useCallback(() => {
-        builder.removeSection(section.id, renameNextUnits.current)
-        refresh()
-    }, [builder, refresh, section])
-
-    return (
-        <Row gutter={MODAL_FORM_EDIT_GUTTER} css={FormModalCss}>
-            <Col span={24}>
-                <Space direction='vertical' size={8}>
-                    <Typography.Text type='secondary'>{NameLabel}</Typography.Text>
-                    <InputNumber
-                        value={name}
-                        min={1}
-                        placeholder={NamePlaceholderLabel}
-                        onChange={setNameValue}
-                        style={INPUT_STYLE}
-                    />
-                </Space>
-                <Checkbox defaultChecked onChange={toggleRenameNextUnits} style={MODAL_FORM_CHECKBOX_STYLE}>
-                    {RenameNextUnitsLabel}
-                </Checkbox>
-            </Col>
-            <Row gutter={MODAL_FORM_BUTTON_GUTTER}>
-                <Col span={24}>
-                    <Button
-                        secondary
-                        onClick={updateSection}
-                        type='sberDefaultGradient'
-                        disabled={isEmpty(name)}
-                    >{SaveLabel}</Button>
-                </Col>
-                <Col span={24}>
-                    <Button
-                        secondary
-                        onClick={deleteSection}
-                        type='sberDangerGhost'
-                        icon={<DeleteFilled />}
-                        style={FULL_SIZE_UNIT_STYLE}
-                    >{DeleteLabel}</Button>
-                </Col>
-            </Row>
-        </Row>
-    )
-}
-
-const TEXT_BUTTON_STYLE: React.CSSProperties = { cursor: 'pointer', marginTop: '8px', display: 'block' }
-
-const AddParkingForm: React.FC<IPropertyMapModalForm> = ({ builder, refresh }) => {
-    const intl = useIntl()
-    const ParkingNameLabel = intl.formatMessage({ id: 'pages.condo.property.parking.form.numberOfParkingSection' })
-    const MinFloorLabel = intl.formatMessage({ id: 'pages.condo.property.section.form.minfloor' })
-    const FloorCountLabel = intl.formatMessage({ id: 'pages.condo.property.parking.form.floorCount' })
-    const ParkingOnFloorLabel = intl.formatMessage({ id: 'pages.condo.property.parking.form.parkingOnFloor' })
-    const ShowMinFloor = intl.formatMessage({ id: 'pages.condo.property.parking.form.showMinFloor' })
-    const HideMinFloor = intl.formatMessage({ id: 'pages.condo.property.parking.form.hideMinFloor' })
-    const AddLabel = intl.formatMessage({ id: 'Add' })
-    const CreateNewLabel = intl.formatMessage({ id: 'pages.condo.property.section.form.mode.create' })
-    const CopyLabel = intl.formatMessage({ id: 'pages.condo.property.parking.form.mode.copy' })
-
-    const [minFloor, setMinFloor] = useState<number>(1)
-    const [floorCount, setFloorCount] = useState<number | null>(null)
-    const [unitsOnFloor, setUnitsOnFloor] = useState(null)
-    const [minFloorHidden, setMinFloorHidden] = useState<boolean>(true)
-    const [copyId, setCopyId] = useState<string | null>(null)
-    const [parkingName, setParkingName] = useState<string>(builder.nextParkingName)
-
-    const toggleMinFloorVisible = useCallback(() => {
-        setMinFloor(1)
-        setMinFloorHidden(!minFloorHidden)
-    }, [minFloorHidden])
-    const setMinFloorValue = useCallback((value) => { setMinFloor(value) }, [])
-    const setFloorCountValue = useCallback((value) => { setFloorCount(value) }, [])
-    const setParkingNameValue = useCallback((value) => setParkingName(value ? value.toString() : ''), [])
-    const maxFloorValue = useMemo(() => {
-        if (floorCount === 1) return minFloor
-        if (minFloor > 0) return floorCount + minFloor - 1
-        return floorCount + minFloor
-    }, [floorCount, minFloor])
-
-    const resetForm = useCallback(() => {
-        setMinFloor(1)
-        setFloorCount(null)
-        setUnitsOnFloor(null)
-    }, [])
-
-    useEffect(() => {
-        if (minFloor && floorCount && unitsOnFloor && parkingName) {
-            builder.addPreviewParking({
-                id: '',
-                name: parkingName,
-                minFloor,
-                maxFloor: maxFloorValue,
-                unitsOnFloor,
-            })
-            refresh()
-        } else {
-            builder.removePreviewParking()
-        }
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [minFloor, floorCount, unitsOnFloor, parkingName])
-
-    useEffect(() => {
-        if (copyId !== null) {
-            builder.addPreviewCopyParking(copyId)
-        }
-    }, [builder, copyId])
-
-    const handleFinish = useCallback(() => {
-        builder.removePreviewParking()
-        if (copyId === null) {
-            builder.addParking({ id: '', name: parkingName, minFloor, maxFloor: maxFloorValue, unitsOnFloor })
-            setParkingName(builder.nextParkingName)
-        } else {
-            builder.addCopyParking(copyId)
-        }
-
-        refresh()
-        resetForm()
-    }, [refresh, resetForm, builder, minFloor, unitsOnFloor, maxFloorValue, parkingName, copyId])
-
-    const isSubmitDisabled = copyId !== null ? false : !(minFloor && floorCount && unitsOnFloor)
-    const isCreateColumnsHidden = copyId !== null
-    const iconRotation = minFloorHidden ? 0 : 180
-    const minFloorMargin = minFloorHidden ? '-28px' : 0
-
-    return (
-        <Row gutter={MODAL_FORM_ROW_GUTTER} css={FormModalCss}>
-            <Col span={24}>
-                <Select value={copyId} onSelect={setCopyId} disabled={builder.isEmptyParking}>
-                    <Select.Option key='create' value={null}>{CreateNewLabel}</Select.Option>
-                    {builder.map.parking.filter(section => !section.preview).map(section => (
-                        <Select.Option
-                            key={`copy-${section.id}`}
-                            value={section.id}
-                        >
-                            {CopyLabel} {section.name}
-                        </Select.Option>
-                    ))}
-                </Select>
-            </Col>
-            <Col span={24} hidden={isCreateColumnsHidden}>
-                <Space direction='vertical' size={8}>
-                    <Typography.Text type='secondary'>{ParkingNameLabel}</Typography.Text>
-                    <InputNumber
-                        value={parkingName}
-                        min={1}
-                        onChange={setParkingNameValue}
-                        style={INPUT_STYLE}
-                        type='number'
-                    />
-                </Space>
-            </Col>
-            <Col span={24} hidden={isCreateColumnsHidden}>
-                <Space direction='vertical' size={8}>
-                    <Typography.Text type='secondary'>{FloorCountLabel}</Typography.Text>
-                    <InputNumber
-                        value={floorCount}
-                        onChange={setFloorCountValue}
-                        style={INPUT_STYLE}
-                        type='number'
-                        min={1}
-                    />
-                </Space>
-            </Col>
-            <Col span={24} hidden={isCreateColumnsHidden} style={{ marginTop: minFloorMargin }}>
-                <Space
-                    direction='vertical'
-                    size={8}
-                    hidden={minFloorHidden}
-                >
-                    <Typography.Text type='secondary'>{MinFloorLabel}</Typography.Text>
-                    <InputNumber
-                        value={minFloor}
-                        onChange={setMinFloorValue}
-                        style={INPUT_STYLE}
-                        type='number'
-                    />
-                </Space>
-                <Typography.Text onClick={toggleMinFloorVisible} style={TEXT_BUTTON_STYLE}>
-                    {minFloorHidden ? ShowMinFloor : HideMinFloor} <DownOutlined rotate={iconRotation} />
-                </Typography.Text>
-            </Col>
-            <Col span={24} hidden={isCreateColumnsHidden}>
-                <Space direction='vertical' size={8}>
-                    <Typography.Text type='secondary'>{ParkingOnFloorLabel}</Typography.Text>
-                    <InputNumber min={1} value={unitsOnFloor} onChange={setUnitsOnFloor} style={INPUT_STYLE} type='number'/>
-                </Space>
-            </Col>
-            <Col span={24}>
-                <Button
-                    key='submit'
-                    secondary
-                    onClick={handleFinish}
-                    type='sberDefaultGradient'
-                    style={MODAL_FORM_BUTTON_STYLE}
-                    disabled={isSubmitDisabled}
-                > {AddLabel} </Button>
-            </Col>
-        </Row>
-    )
-}
-
-const EditParkingForm: React.FC<IPropertyMapModalForm> = ({ builder, refresh }) => {
-    const intl = useIntl()
-    const NameLabel = intl.formatMessage({ id: 'pages.condo.property.parking.form.numberOfParkingSection' })
-    const SaveLabel = intl.formatMessage({ id: 'Save' })
-    const DeleteLabel = intl.formatMessage({ id: 'Delete' })
-    const RenameNextParkingUnitsLabel = intl.formatMessage({ id: 'pages.condo.property.modal.RenameNextParkingUnits' })
-
-    const [parkingName, setParkingName] = useState<string>('')
-    const renameNextUnits = useRef(true)
-
-    const parkingSection = builder.getSelectedParking()
-
-    const deleteParking = useCallback(() => {
-        builder.removeParking(parkingSection.id, renameNextUnits.current)
-        refresh()
-    }, [builder, refresh, parkingSection])
-
-    const updateParkingSection = useCallback(() => {
-        builder.updateParking({ ...parkingSection, name: parkingName })
-        refresh()
-    }, [builder, refresh, parkingName, parkingSection])
-
-    const setParkingNameValue = useCallback((value) => setParkingName(value ? value.toString() : ''), [])
-    const toggleRenameNextUnits = useCallback((event) => { renameNextUnits.current = event.target.checked }, [])
-
-    useEffect(() => {
-        setParkingName(parkingSection ? parkingSection.name : '')
-    }, [parkingSection])
-
-    return (
-        <Row gutter={MODAL_FORM_EDIT_GUTTER} css={FormModalCss}>
-            <Col span={24}>
-                <Space direction='vertical' size={8}>
-                    <Typography.Text type='secondary'>{NameLabel}</Typography.Text>
-                    <InputNumber
-                        value={parkingName}
-                        min={1}
-                        onChange={setParkingNameValue}
-                        style={INPUT_STYLE}
-                    />
-                </Space>
-                <Checkbox defaultChecked onChange={toggleRenameNextUnits} style={MODAL_FORM_CHECKBOX_STYLE}>
-                    {RenameNextParkingUnitsLabel}
-                </Checkbox>
-            </Col>
-            <Row gutter={MODAL_FORM_BUTTON_GUTTER}>
-                <Col span={24}>
-                    <Button
-                        secondary
-                        onClick={updateParkingSection}
-                        type='sberDefaultGradient'
-                    >
-                        {SaveLabel}
-                    </Button>
-                </Col>
-                <Col span={24}>
-                    <Button
-                        secondary
-                        onClick={deleteParking}
-                        type='sberDangerGhost'
-                        icon={<DeleteFilled />}
-                        style={FULL_SIZE_UNIT_STYLE}
-                    >
-                        {DeleteLabel}
-                    </Button>
-                </Col>
-            </Row>
-        </Row>
-    )
-}
-
-const ParkingUnitForm: React.FC<IPropertyMapModalForm> = ({ builder, refresh }) => {
-    const intl = useIntl()
-    const mode = builder.editMode
-    const SaveLabel = intl.formatMessage({ id: mode === 'editParkingUnit' ? 'Save' : 'Add' })
-    const DeleteLabel = intl.formatMessage({ id: 'Delete' })
-    const NameLabel = intl.formatMessage({ id: 'pages.condo.property.parkingUnit.Name' })
-    const SectionLabel = intl.formatMessage({ id: 'pages.condo.property.parkingSection.name' })
-    const FloorLabel = intl.formatMessage({ id: 'pages.condo.property.floor.Name' })
-    const SectionTitlePrefix = intl.formatMessage({ id: 'pages.condo.property.select.option.parking' })
-    const RenameNextParkingUnitsLabel = intl.formatMessage({ id: 'pages.condo.property.modal.RenameNextParkingUnits' })
-
-    const [label, setLabel] = useState('')
-    const [floor, setFloor] = useState('')
-    const [section, setSection] = useState('')
-
-    const [sections, setSections] = useState([])
-    const [floors, setFloors] = useState([])
-
-    const renameNextUnits = useRef(true)
-
-    const updateSection = (value) => {
-        setSection(value)
-        setFloors(builder.getParkingSectionFloorOptions(value))
-        if (mode === 'editParkingUnit') {
-            const mapUnit = builder.getSelectedParkingUnit()
-            if (value === mapUnit.section) {
-                setFloor(mapUnit.floor)
-            } else {
-                setFloor(null)
-            }
-        } else {
-            setFloor(null)
-        }
-    }
-
-    useEffect(() => {
-        setSections(builder.getParkingSectionOptions())
-        const mapUnit = builder.getSelectedParkingUnit()
-        if (mapUnit) {
-            setFloors(builder.getParkingSectionFloorOptions(mapUnit.section))
-            setLabel(mapUnit.label)
-            setSection(mapUnit.section)
-            setFloor(mapUnit.floor)
-        }
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [builder])
-
-    useEffect(() => {
-        if (label && floor && section && mode === 'addParkingUnit') {
-            builder.addPreviewParkingUnit({ id: '', label, floor, section }, renameNextUnits.current)
-            refresh()
-        } else {
-            builder.removePreviewParkingUnit(renameNextUnits.current)
-        }
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [label, floor, section, mode])
-
-    const resetForm = useCallback(() => {
-        setLabel('')
-        setFloor('')
-        setSection('')
-    }, [])
-
-    const toggleRenameNextUnits = useCallback((event) => { renameNextUnits.current = event.target.checked }, [])
-
-    const applyChanges = useCallback(() => {
-        const mapUnit = builder.getSelectedParkingUnit()
-        if (mapUnit) {
-            builder.updateParkingUnit({ ...mapUnit, label, floor, section }, renameNextUnits.current)
-        } else {
-            builder.removePreviewParkingUnit()
-            builder.addParkingUnit({ id: '', label, floor, section }, renameNextUnits.current)
-            resetForm()
-        }
-        refresh()
-    }, [builder, refresh, resetForm, label, floor, section])
-
-    const deleteUnit = useCallback(() => {
-        const mapUnit = builder.getSelectedParkingUnit()
-        builder.removeParkingUnit(mapUnit.id, renameNextUnits.current)
-        refresh()
-        resetForm()
-    }, [resetForm, refresh, builder])
-
-    const isApplyButtonDisabled = useMemo(() => {
-        let isUnitLabelUnique = true
-        if (mode === 'addParkingUnit') {
-            isUnitLabelUnique = builder.validateUniqueUnitLabel()
-        } else if (mode === 'editParkingUnit') {
-            const selectedUnit = builder.getSelectedParkingUnit()
-            const unitPlacementChanged = selectedUnit.floor !== floor || selectedUnit.section !== section
-            const labelChanged = selectedUnit.label !== label
-            const labelValidation = labelChanged
-                ? builder.validateUniqueUnitLabel(label, 'parking')
-                : false
-
-            isUnitLabelUnique = unitPlacementChanged
-                ? (!labelChanged || labelValidation)
-                : labelValidation
-        }
-
-        return !(floor && section && label.trim() && isUnitLabelUnique)
-    }, [floor, section, label, builder, mode])
-
-    return (
-        <Row gutter={MODAL_FORM_ROW_GUTTER} css={FormModalCss}>
-            <Col span={24}>
-                <Space direction='vertical' size={8} style={INPUT_STYLE}>
-                    <Typography.Text type='secondary' >{SectionLabel}</Typography.Text>
-                    <Select value={section} onSelect={updateSection} style={INPUT_STYLE}>
-                        {sections.map((sec) => {
-                            return <Option key={sec.id} value={sec.id}>{SectionTitlePrefix} {sec.label}</Option>
-                        })}
-                    </Select>
-                </Space>
-            </Col>
-            <Col span={24}>
-                <Space direction='vertical' size={8} style={INPUT_STYLE}>
-                    <Typography.Text type='secondary' >{FloorLabel}</Typography.Text>
-                    <Select value={floor} onSelect={setFloor} style={INPUT_STYLE}>
-                        {floors.map(floorOption => {
-                            return <Option key={floorOption.id} value={floorOption.id}>{floorOption.label}</Option>
-                        })}
-                    </Select>
-                </Space>
-            </Col>
-            <Col span={24}>
-                <Space direction='vertical' size={BUTTON_SPACE_SIZE}>
-                    <Space direction='vertical' size={8}>
-                        <Typography.Text type='secondary'>{NameLabel}</Typography.Text>
-                        <Input allowClear={true} value={label} onChange={e => setLabel(e.target.value)} style={INPUT_STYLE} />
-                        <Checkbox defaultChecked onChange={toggleRenameNextUnits} style={MODAL_FORM_CHECKBOX_STYLE}>
-                            {RenameNextParkingUnitsLabel}
-                        </Checkbox>
-                    </Space>
-                    <Row gutter={MODAL_FORM_ROW_BUTTONS_GUTTER}>
-                        <Col span={24}>
-                            <Button
-                                secondary
-                                onClick={applyChanges}
-                                type='sberDefaultGradient'
-                                disabled={isApplyButtonDisabled}
-                            > {SaveLabel} </Button>
-                        </Col>
-                        {
-                            mode === 'editParkingUnit' && (
-                                <Col span={24}>
-                                    <Button
-                                        secondary
-                                        onClick={deleteUnit}
-                                        type='sberDangerGhost'
-                                        icon={<DeleteFilled />}
-                                    >{DeleteLabel}</Button>
-                                </Col>
-                            )
-                        }
-                    </Row>
-                </Space>
-            </Col>
-        </Row>
-    )
-}
-
-const AddSectionFloor: React.FC<IPropertyMapModalForm> = ({ builder, refresh }) => {
-    const intl = useIntl()
-    const SaveLabel = intl.formatMessage({ id: 'Add' })
-    const UnitTypeAtFloorLabel = intl.formatMessage({ id: 'pages.condo.property.modal.title.unitTypeAtFloor' })
-    const UnitsOnFloorLabel = intl.formatMessage({ id: 'pages.condo.property.section.form.unitsOnFloor' })
-    const SectionLabel = intl.formatMessage({ id: 'pages.condo.property.parkingSection.name' })
-    const FloorLabel = intl.formatMessage({ id: 'pages.condo.property.floor.Name' })
-    const SectionTitlePrefix = intl.formatMessage({ id: 'pages.condo.property.select.option.section' })
-    const RenameNextUnitsLabel = intl.formatMessage({ id: 'pages.condo.property.modal.RenameNextUnits' })
-
-    const [sections, setSections] = useState([])
-    const [section, setSection] = useState<number | null>(null)
-    const [unitType, setUnitType] = useState<BuildingUnitSubType>(BuildingUnitSubType.Flat)
-    const [unitsOnFloor, setUnitsOnFloor] = useState<number>()
-    const [floor, setFloor] = useState<string>('')
-
-    const maxFloor = useRef<number>(0)
-    const renameNextUnits = useRef(true)
-
-    const setFloorNumber = useCallback((value) => setFloor(value ? value.toString() : ''), [])
-    const setUnitsOnFloorNumber = useCallback((value) => setUnitsOnFloor(value ? value.toString() : ''), [])
-    const toggleRenameNextUnits = useCallback((event) => { renameNextUnits.current = event.target.checked }, [])
-    const applyChanges = useCallback(() => {
-        if (floor !== '' && section !== null && unitsOnFloor > 0) {
-            builder.addSectionFloor({
-                section: Number(section),
-                index: Number(floor),
-                unitType,
-                unitCount: Number(unitsOnFloor),
-            }, renameNextUnits.current)
-        }
-        refresh()
-
-        setUnitsOnFloor(null)
-        setFloor(null)
-        setSection(null)
-    }, [builder, refresh, floor, section, unitsOnFloor, unitType])
-
-    const isSubmitDisabled = useMemo(() => {
-        const sectionFloors = builder.getSectionFloorNames(section)
-        return !(floor && section !== null && unitsOnFloor && !sectionFloors.includes(floor))
-    }, [floor, section, unitsOnFloor])
-
-    useEffect(() => {
-        if (section !== null) {
-            maxFloor.current = builder.getSectionMaxFloor(section) + 1
-        }
-    }, [section])
-
-    useEffect(() => {
-        setSections(builder.getSectionOptions())
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [builder])
-
-    useEffect(() => {
-        if (floor && section !== null && unitsOnFloor > 0) {
-            const sectionFloors = builder.getSectionFloorNames(section)
-            if (!sectionFloors.includes(floor)) {
-                builder.addPreviewSectionFloor({
-                    section: Number(section),
-                    index: Number(floor),
-                    unitType,
-                    unitCount: Number(unitsOnFloor),
-                })
-            }
-        } else {
-            builder.removePreviewSectionFloor()
-        }
-        refresh()
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [floor, section, unitsOnFloor, unitType])
-
-
-
-    return (
-        <Row gutter={MODAL_FORM_ROW_GUTTER} css={FormModalCss}>
-            <Col span={24}>
-                <Space direction='vertical' size={8} style={INPUT_STYLE}>
-                    <Typography.Text type='secondary'>{UnitTypeAtFloorLabel}</Typography.Text>
-                    <Select value={unitType} onSelect={setUnitType}>
-                        {Object.values(BuildingUnitSubType)
-                            .filter(unitType => unitType !== BuildingUnitSubType.Parking)
-                            .map((unitType, key) => (
-                                <Select.Option key={`${key}-${unitType}`} value={unitType} title={unitType}>
-                                    {intl.formatMessage({ id: `pages.condo.property.modal.unitType.${unitType}` })}
-                                </Select.Option>
-                            ))
-                        }
-                    </Select>
-                </Space>
-            </Col>
-            <Col span={24}>
-                <Space direction='vertical' size={8} style={INPUT_STYLE}>
-                    <Typography.Text type='secondary'>{SectionLabel}</Typography.Text>
-                    <Select value={section} onSelect={setSection} style={INPUT_STYLE}>
-                        {sections.map((sec, index) => {
-                            return <Option key={sec.id} value={index}>{SectionTitlePrefix} {sec.label}</Option>
-                        })}
-                    </Select>
-                </Space>
-            </Col>
-            <Col span={24}>
-                <Space direction='vertical' size={8} style={INPUT_STYLE}>
-                    <Typography.Text type='secondary'>{FloorLabel}</Typography.Text>
-                    <InputNumber
-                        value={floor}
-                        onChange={setFloorNumber}
-                        max={maxFloor.current}
-                        type='number'
-                        style={INPUT_STYLE}
-                    />
-                </Space>
-            </Col>
-            <Col span={24}>
-                <Space direction='vertical' size={BUTTON_SPACE_SIZE}>
-                    <Space direction='vertical' size={8}>
-                        <Typography.Text type='secondary'>{UnitsOnFloorLabel}</Typography.Text>
-                        <InputNumber
-                            value={unitsOnFloor}
-                            onChange={setUnitsOnFloorNumber}
-                            style={INPUT_STYLE}
-                            type='number'
-                            min={1}
-                        />
-                        <Checkbox defaultChecked onChange={toggleRenameNextUnits} style={MODAL_FORM_CHECKBOX_STYLE}>
-                            {RenameNextUnitsLabel}
-                        </Checkbox>
-                    </Space>
-                    <Row gutter={MODAL_FORM_ROW_BUTTONS_GUTTER}>
-                        <Col span={24}>
-                            <Button
-                                secondary
-                                onClick={applyChanges}
-                                type='sberDefaultGradient'
-                                disabled={isSubmitDisabled}
-                            > {SaveLabel} </Button>
-                        </Col>
-                    </Row>
-                </Space>
-            </Col>
-        </Row>
-    )
-}
-
